Выносит интерфейсы в отдельный модуль ports.ts
