+++ src/js/types.ts
@@ -1,7 +1,3 @@
-import { Order, Product, ProductId, SentOrder } from './models';
-
-export * from './models';
-
 export type AppEvents =
 	| 'START'
 	| 'CARD_SELECT'
@@ -17,32 +13,3 @@ export type AppEvents =
 	| 'ORDER_CONTACT_SUBMIT'
 	| 'SUCCESS_CLOSE';
 
-// ~~~~~~~~~~~ Порт "Services" ~~~~~~~~~~~ //
-
-export interface IProductService {
-	getProducts: () => Promise<Product[]>;
-	getProduct: (id: ProductId) => Promise<Product>;
-}
-
-export interface IOrderService {
-	sendOrder: (order: Order) => Promise<SentOrder>;
-}
-
-export interface IWebLarekApi {
-	getProducts: () => Promise<Product[]>;
-	getProduct: (id: ProductId) => Promise<Product>;
-	postOrder: (order: Order) => Promise<SentOrder>;
-}
-
-// ~~~~~~~~~~~~~ Порт "View" ~~~~~~~~~~~~~ //
-
-export interface IModalView<H> {
-	content: H;
-	render: (data: { content: H }) => H;
-	open: () => void;
-	close: () => void;
-}
-
-export interface IAppView<T, H> {
-	render: (data?: Partial<T>) => H;
-}
